LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.8_Literals/7.8.5_Regular_Expression_Literals/S7.8.5_A2.2_T2-expected.txt

CONSOLE MESSAGE: line 77: SyntaxError: Parse error
S7.8.5_A2.2_T2

PASS Expected parsing failure

TEST COMPLETE
